app/requirements.txt
django==2.1.3
django-cors-headers==2.4.0
djangorestframework==3.9.0
#mysqlclient==1.3.13
django-filter==2.0.0
django-mdeditor==0.1.9
django-redis==4.9.0
redis==2.10.6
requests==2.20.0
uwsgi==2.0.17.1
